pyproject.toml
[project]
name = "nostrike"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
dependencies = [
    "typer",
    "mpire",
    "dnspython",
    "pyfiglet",
    "nuitka",
    "pytest",
]

[tool.uv]
dev-dependencies = []
